+++ pvoutput/grid_search/app.py
@@ -2,8 +2,9 @@ import argparse
 import logging
 import os
 import sys
+from typing import Optional
 
-from pvoutput.grid_search.grid_search import query_yes_no, GridSearch
+from pvoutput.grid_search.grid_search import GridSearch
 from pvoutput.grid_search.natural_earth import NaturalEarth
 
 
@@ -178,3 +179,36 @@ if __name__ == "__main__":
     datefmt = "%Y-%m-%dT%H:%M:%SZ"
     logging.basicConfig(format=fmt, datefmt=datefmt, level=os.environ.get("LOGLEVEL", "INFO"))
     main()
+
+
+def query_yes_no(question: str, default: Optional[str] = "yes") -> bool:
+    """Ask a yes/no question via input() and return the answer as boolean.
+
+    Args:
+        question:
+            The question presented to the user.
+        default:
+            The presumed answer if the user just hits <Enter>. It must be "yes" (the default), "no"
+            or None (meaning an answer is required of the user).
+
+    Returns:
+        Return value is True for "yes" or False for "no".
+    """
+    valid = {"yes": True, "y": True, "ye": True, "no": False, "n": False}
+    if default is None:
+        prompt = " [y/n] "
+    elif default == "yes":
+        prompt = " [Y/n] "
+    elif default == "no":
+        prompt = " [y/N] "
+    else:
+        raise ValueError("invalid default answer: '%s'" % default)
+    while True:
+        sys.stdout.write(question + prompt)
+        choice = input().lower()
+        if default is not None and choice == "":
+            return valid[default]
+        elif choice in valid:
+            return valid[choice]
+        else:
+            sys.stdout.write("Please respond with 'yes' or 'no' " "(or 'y' or 'n').\n")
