+++ pvoutput/grid_search/grid_search.py
@@ -13,9 +13,6 @@ Provides both an importable method and a CLI.
 - First Authored: 2021-11-16
 """
 
-import logging
-import os
-import sys
 from typing import Iterable, Optional, Tuple, Union
 
 import geopandas as gpd
@@ -25,7 +22,6 @@ import pandas as pd
 from pyproj import Transformer
 from shapely.geometry import Point, Polygon
 
-from pvoutput.grid_search.app import main
 from pvoutput.grid_search.natural_earth import NaturalEarth
 
 
@@ -440,34 +436,3 @@ def clip_to_radius(
     return coords.loc[coords["selected"]].drop(columns="selected")
 
 
-def query_yes_no(question: str, default: Optional[str] = "yes") -> bool:
-    """Ask a yes/no question via input() and return the answer as boolean.
-
-    Args:
-        question:
-            The question presented to the user.
-        default:
-            The presumed answer if the user just hits <Enter>. It must be "yes" (the default), "no"
-            or None (meaning an answer is required of the user).
-
-    Returns:
-        Return value is True for "yes" or False for "no".
-    """
-    valid = {"yes": True, "y": True, "ye": True, "no": False, "n": False}
-    if default is None:
-        prompt = " [y/n] "
-    elif default == "yes":
-        prompt = " [Y/n] "
-    elif default == "no":
-        prompt = " [y/N] "
-    else:
-        raise ValueError("invalid default answer: '%s'" % default)
-    while True:
-        sys.stdout.write(question + prompt)
-        choice = input().lower()
-        if default is not None and choice == "":
-            return valid[default]
-        elif choice in valid:
-            return valid[choice]
-        else:
-            sys.stdout.write("Please respond with 'yes' or 'no' " "(or 'y' or 'n').\n")
